+++ app/admin/dashboard/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const dynamic = 'force-dynamic'
+
+export default async function User() {
+  return <div className="container p-5"></div>
+}
