+++ app/admin/layout.tsx
@@ -1,0 +1,45 @@
+import { createServerComponentClient } from '@supabase/auth-helpers-nextjs'
+import { redirect } from 'next/navigation'
+import { cookies } from 'next/headers'
+import { AuthButtonSignOut } from '@/components'
+import { Avatar } from '@radix-ui/themes'
+
+export const dynamic = 'force-dynamic'
+
+export const metadata = {
+  title: 'DivSplash',
+  description: 'Welcome to DivSplash Babyyy!',
+}
+
+export default async function RootLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  const supabase = createServerComponentClient<Database>({ cookies })
+
+  const {
+    data: { session },
+  } = await supabase.auth.getSession()
+
+  if (!session) {
+    redirect('/')
+  }
+
+  const { data: user } = await supabase.from('profiles').select('*').single()
+
+  console.log(user)
+
+  if (user?.role !== 'admin') {
+    return redirect('/')
+  }
+
+  return (
+    <main className="container p-5 space-y-5">
+      <h1 className="text-lg font-bold text-center lg:text-left">
+        DivSplash Admin Panel 💻
+      </h1>
+      {children}
+    </main>
+  )
+}
