+++ app/not-found.tsx
@@ -1,0 +1,7 @@
+export default function NotFoundPage() {
+  return (
+    <h1 className="mt-10 text-5xl font-bold text-center text-rose-500">
+      Page not found
+    </h1>
+  )
+}
